+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support `select` element (it was blocked by a built-in Ruby method)
 - Handle case of `:parent` selector being invalid, defaulting to `body`.
 - Remove pure-struct gem dependency as the latest Opal fixed the implementation of Struct
+- New Hello, Data-Binding! Sample: `require 'glimmer-dsl-web/samples/hello/hello_data_binding'`
 
 ## 0.0.5